Allow a define() method in the input file.
In that case, it will be called to return the input definitions.

The job of the define() function is to return a dictionary of parameters. How
the dictionary is constructed is not our business, but the usual way is to
simply have a function define() along these lines in the input file:

    def define():
        options = {
            'saveEigVectors' : None,
            'eigenSolver' : 'eigen1',
        }
        region_2 = {
            'name' : 'Surface',
            'select' : 'nodes of surface',
        }
        ...
        return locals()